refactored the code to use vertical slice architecture
