+++ .gitignore
@@ -1,0 +1,3 @@
+.idea
+.DS_Store
+.vs
