+++ internal/features/members/models.go
@@ -1,0 +1,10 @@
+package members
+
+// Member represents a gym member
+type Member struct {
+	ID         int    `json:"id"`
+	Name       string `json:"name"`
+	Email      string `json:"email"`
+	Age        int    `json:"age"`
+	JoinedDate string `json:"joined_date"`
+}
